fix: confidential app client should also not be used as api access client
Basically we have 2 authentication flows:

1. frontend (i.e. user) requests an public user token to auth server, exchange that token to an confidential user token through backend route exchange-token and then use that token for each subsequent call to backend

2. an user or machine (scripts, jupyter notebooks, etc...) wants to connect to the backend api, it needs to request an api access token through backend route issue-api-token using his api client secret key (think of it as an api key) and then use that token for each subsequent call to backend

Note: Authentication flow #2 is machete (gambiarra, kludge, τσαπατσουλιά /tsapatsoulia/) because the auth should link an api key for each user instead of "linking" multiple users to an application through e-groups. That is, for fine-grained api access roles instead of given a role to each user we need to create an application to each user (this is inefficient because will generate multiple applications) or group multiple users in an application (this is inefficient because multiple users would share the same secret key).